fix: move order related entities to api package
