+++ ftgo-order-service-api/src/main/kotlin/me/jangjunha/ftgo/order_service/api/RevisedOrderLineItem.kt
@@ -1,4 +1,4 @@
-package me.jangjunha.ftgo.order_service.domain
+package me.jangjunha.ftgo.order_service.api
 
 data class RevisedOrderLineItem(
     val quantity: Int,
